success to auto register and detect

+++ pkg/config.go
@@ -13,7 +13,7 @@ type Config struct {
 	//FaceSetName      string
 	DBConnStr     string
 	Diskthreshold int
-	Similarity    int64
+	Similarity    int
 }
 
 func InitConfig() *Config {
@@ -26,7 +26,7 @@ func InitConfig() *Config {
 	//flag.StringVar(&config.FaceSetName, "faceset-name", "", "faceset to work with")
 	flag.StringVar(&config.DBConnStr, "db-conn-str", "", "db connection string (user:password@tcp(ip:port)/dbname)")
 	flag.IntVar(&config.Diskthreshold, "disk-threshod", 80, "images data store in disk to trigger to rm images")
-	flag.Int64Var(&config.Similarity, "similarity", 92, "similarity to judge two image are similar")
+	flag.IntVar(&config.Similarity, "similarity", 92, "similarity to judge two image are similar")
 
 	flag.Parse()
 	return &config
